// src/app/common/interfaces/scheme.interface.ts
export interface SchemeVsBudget {
    schemeno: string;
    description: string;
    retailercode: string;
    customername: string;
    budgetamount: number;
    budgetquantity: number;
}



export interface CustomerWiseMaxQuantity{
    parac:string;
    acname: any;
    mcode:string;
    desca: any;
    maxquantity:number;
}